# server/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RunMe API",
    "description": "REST API for the RunMe mobile app (auth, loans, grants, wallet, admin). Use **Authorize** with a Bearer token from `POST /auth/login` or `POST /auth/register`.",
    "version": "0.1.0"
  },
  "servers": [
    { "url": "/", "description": "Current host" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the `accessToken` value from login/register (without the word Bearer)."
      }
    },
    "schemas": {
...
      "Loan": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "borrowerId": { "type": "string", "format": "uuid" },
          "lenderId": { "type": "string", "format": "uuid", "nullable": true },
          "amount": { "type": "number" },
          "purpose": { "type": "string", "enum": ["rent", "food", "emergency"] },
          "durationDays": { "type": "integer" },
          "interestRate": { "type": "number" },
          "status": { "type": "string", "enum": ["pending", "funded", "repaid"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "audience": { "type": "string", "enum": ["public", "friendsOnly"] },
          "reason": { "type": "string", "nullable": true },
          "proofFileUrl": { "type": "string", "nullable": true },
          "repaidAmount": { "type": "number" }
        }
      },
      "LoanCreate": {
        "type": "object",
        "required": ["amount", "purpose", "durationDays", "audience"],
        "properties": {
          "amount": { "type": "number", "minimum": 0 },
          "purpose": { "type": "string", "enum": ["rent", "food", "emergency"] },
          "durationDays": { "type": "integer", "minimum": 7, "maximum": 365 },
          "audience": { "type": "string", "enum": ["public", "friendsOnly"] },
          "reason": { "type": "string" },
          "proofFileUrl": { "type": "string", "format": "uri", "nullable": true }
        }
      },
      "RepaymentRequest": {
        "type": "object",
        "required": ["amount"],
        "properties": {
          "amount": { "type": "number", "minimum": 0 }
        }
      },
      "Backing": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "loanId": { "type": "string", "format": "uuid" },
          "backerId": { "type": "string", "format": "uuid" },
          "amountGuaranteed": { "type": "number" }
        }
      },
      "BackingCreate": {
        "type": "object",
        "required": ["amountGuaranteed"],
        "properties": {
          "amountGuaranteed": { "type": "number", "minimum": 0 }
        }
      },
      "Coverage": {
        "type": "object",
        "properties": {
          "principal": { "type": "number" },
          "totalGuaranteed": { "type": "number" },
          "coverageRatio": { "type": "number" }
        }
      },
...
    }
  },
  "paths": {
...
    "/admin/loans": {
      "get": {
        "tags": ["Admin"],
        "summary": "List all loans",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
    "/loans": {
      "get": {
        "tags": ["Loans"],
        "summary": "List loans",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "borrowerId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "funded", "repaid"] } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/Loan" } },
                    "nextCursor": { "type": "string", "nullable": true }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["Loans"],
        "summary": "Create loan request",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoanCreate" } }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Loan" } } }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/loans/{loanId}": {
      "get": {
        "tags": ["Loans"],
        "summary": "Get loan by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "loanId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Loan" } } }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/loans/{loanId}/coverage": {
      "get": {
        "tags": ["Loans"],
        "summary": "Funding coverage",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "loanId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Coverage" } } }
          },
          "404": { "description": "Not found" }
        }
      }
    },
    "/loans/{loanId}/backings": {
      "get": {
        "tags": ["Loans"],
        "summary": "List backings",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "loanId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/Backing" } }
                  }
                }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      },
      "post": {
        "tags": ["Loans"],
        "summary": "Create backing / guarantee",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "loanId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/BackingCreate" } }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Backing" } } }
          },
          "404": { "description": "Not found" },
          "409": { "description": "Conflict" }
        }
      }
    },
    "/loans/{loanId}/repayments": {
      "post": {
        "tags": ["Loans"],
        "summary": "Repay loan",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "loanId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RepaymentRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Loan" } } }
          },
          "400": { "description": "Bad request" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" },
          "409": { "description": "Wrong state" }
        }
      }
    },
...
  }
}